added fuse.py for general setup functions, keeping gdal.py specifc (as in 0.3.x series).
